Autoboxing Integers

+++ ch04/ch04_01/TestLinkedList.java
@@ -4,10 +4,10 @@ public class TestLinkedList
 {
 	public static void main( String[] args )
 	{
-		LinkedList<Double> LL = new LinkedList<Double>();
+		LinkedList<Integer> LL = new LinkedList<Integer>();
 		
-		LL.insertAtHead(5.0);
-		LL.insertAtHead(10.0);
+		LL.insertAtHead(5);
+		LL.insertAtHead(10);
 		/*LL.insertAtHead(2);
 		LL.insertAtHead(12);
 		LL.insertAtHead(19);
@@ -19,7 +19,7 @@ public class TestLinkedList
 		System.out.println(LL);
 		System.out.println("Length : " + LL.length());
 		
-		System.out.println("Find 12 ? : " + LL.find(12.0));
+		System.out.println("Find 12 ? : " + LL.find(12));
 		//System.out.println("Find 77 ? : " + LL.find(77));
 	}
 }
